ADM update
Includes the function to create a new item for an ADM user. Also fiz the bug of login/logout and change the color of logout sign. Change the structure of user: now all user are loaded in the begining of the executation. Fix the show clients function. Adds a new client called "generic" for tests (only).