Creation of a client user
Delete the standart library constructor. Add a functional method for create a new client user (open for any user), with a email existence detection. Add a wait to main.cpp and change the application to a pointer in main. Coment in takeSingleInfo is updated. Update the load client items method. Also updates the show client items method, to include the possibility of an empty library. An option 'forgot my password' was added.